added weight-list

+++ src/main/java/app/redoge/yhshback/dto/UserUpdateRequestDto.java
@@ -1,6 +1,6 @@
 package app.redoge.yhshback.dto;
 
 
-public record UserUpdateRequestDto(String username, String sex, int weightKg, int heightSm) {
+public record UserUpdateRequestDto(String username, String sex, float weightKg, int heightSm) {
 
 }
